fix: add fallback for --save-hf when Megatron-Bridge lacks model support
When Megatron-Bridge doesn't support a model architecture, save_hf_model
now falls back to reading the just-saved torch_dist checkpoint and
converting to HF format via the existing conversion logic.

Extract shared checkpoint loading/saving utilities from
tools/convert_torch_dist_to_hf.py into slime/utils/torch_dist_to_hf.py
so both the CLI tool and the runtime fallback can reuse them.
